net/bnxt: tf_ulp: TFC support flow scale query for Thor2
TFC supports the flow scale query feature for OVS application.
The resource usage(WC-TCAM) is tracked inside Thor2 firmware.
This patch is to query the wc-tcam usage info when adding/
deleting a flow. It is just for debugging purpose and disabled by
default.

Using the build flag -DTF_FLOW_SCALE_QUERY to enable it.
And users should use niccli to query the resource usage.

Signed-off-by: Shuanglin Wang <shuanglin.wang@broadcom.com>
Signed-off-by: Sriharsha Basavapatna <sriharsha.basavapatna@broadcom.com>
Reviewed-by: Kishore Padmanabha <kishore.padmanabha@broadcom.com>
Reviewed-by: Ajit Khaparde <ajit.khaparde@broadcom.com>