net/ice: support custom search path for DDP package files
Add support for customizing the firmware search path for DDP package
files, to match the kernel driver behavior. The PMD will read the search
path from "/sys/module/firmware_class/parameters/path", and try to load
a DDP package from that location before checking the standard package
paths.

When documenting this new behaviour, update the documentation in general
for loading the DDP package, moving it out of the "limitations" section
of the document.

Signed-off-by: Zhichao Zeng <zhichaox.zeng@intel.com>
Acked-by: Bruce Richardson <bruce.richardson@intel.com>